+++ CHANGELOG.md
@@ -22,6 +22,11 @@ corresponding version of the Dune core modules.
 
   To ensure a smooth transition we suggest to not rely on default
   index merging strategies until 2.11 has been released.
+- Similarly, when calling `functions.Power` to construct a product
+  of finite element bases in a Python program, you will now get a warning
+  saying that the default for the `layout` parameter will change
+  from `lexicographic` to `interleaved` after the 2.11 release.
+  After these changes, C++ and Python will have the same defaults.
 - The `LagrangeDG(Pre)Basis` supports selecting the polynomial order at runtime.
   Its static variables `LagrangeDGPreBasis::dofsPer.*` are deprecated.
 - Add function wrappers `CoarseFunctionOnFineGridView` and `FineFunctionOnCoarseGridView`